Migrate Number from GasPrice to its own component
This breaks the ability for the bottom number to drive the top
